Splitting the DB code and adding testing
