+++ SpiNNaker-storage/src/test/resources/log4j.properties
@@ -1,0 +1,9 @@
+# Root logger option
+log4j.rootLogger=INFO, stderr
+log4j.logger.org.apache.commons.beanutils=ERROR
+
+# Direct log messages to stderr
+log4j.appender.stderr=org.apache.log4j.ConsoleAppender
+log4j.appender.stderr.Target=System.err
+log4j.appender.stderr.layout=org.apache.log4j.PatternLayout
+log4j.appender.stderr.layout.ConversionPattern=%d{HH:mm:ss} %-5p %c{1}:%L - %m%n
